### 2016-08-08 #68 @cvitolo
rdefra package
The rdefra package retrieves air pollution data and metadata from the Air Information Resource (UK-AIR) of the Department for Environment, Food and Rural Affairs in the United Kingdom. UK-AIR does not provide a public API for programmatic access to data, therefore this package scrapes the HTML pages to get relevant information.

```
Package: rdefra
Title: Interact with the UK AIR Pollution Database from DEFRA
Version: 0.2.0
Author: Claudia Vitolo [aut, cre], Andrew Russell [aut], Allan Tucker [aut]
Maintainer: Claudia Vitolo <cvitolodev@gmail.com>
Description: Get data from DEFRA's UK-AIR website. It basically scraps the HTML content.
Imports: RCurl, XML, plyr
Suggests: testthat
Repository: CRAN
```

https://github.com/kehraProject/r_rdefra

Scientists and researchers interested in air pollution data and epidemiologists.

The openair package (https://github.com/davidcarslaw/openair) accomplishes similar things but relies on a local and compressed copy of the data on servers at King's College (UK), periodically updated. I have used the openair package myself in the past and it is an excellent package (for data retrieval and visualisation) but I had troubles with King's College servers down time.
The rdefra package, instead, retrieves the information directly from the original source with the advantage that users always get the most complete information at any time. This package also integrates a function to retrieve missing coordinates from the standard metadata.

    - [x] The package contains a [`paper.md`](http://joss.theoj.org/about#paper_structure) with a high-level description.
    http://dx.doi.org/10.5281/zenodo.59851

R CMD check succeeds. There were no ERRORs, WARNINGs or NOTEs.
