### 2016-08-08 #68 @sckott
### Editor checks:

- [x] **Fit**: The package meets or criteria [fit](policies.md#fit) and [overlap](policies.md#fit)
- [x] **Automated tests:** __Package has some tests, but none test actual functions in the package__

---

@cvitolo Thanks for your submission!  Two things before I get reviewers:

- you do have a tests directory, but AFAICT they don't test the functions in the package. Can you add at least a few tests that test the functionality of the exported functions?
- You did check that you have a vignette, but the vignette is in the parent directory, so not included with the package. Please move the vignette dir into the package itself
