### 2016-08-09 #68 @cvitolo
Hi Scott,

Many thanks for looking at my package.
I have made the changes you mentioned below.

Kind regards,
Claudia

On 9 August 2016 at 00:00, Scott Chamberlain <notifications@github.com>
wrote:

> Editor checks:
>
>    - *Fit*: The package meets or criteria fit <http://policies.md#fit>
>    and overlap <http://policies.md#fit>
>    - *Automated tests:* *Package has some tests, but none test actual
>    functions in the package*
>    - *License:* The package has a CRAN or OSI accepted license
>    - *Repository:* The repository link resolves correctly
>    - *Archive* (JOSS only, may be post-review): The repository DOI
>    resolves correctly
>    - *Version* (JOSS only, may be post-review): Does the release version
>    given match the GitHub release (v1.0.0)?
>
> ------------------------------
>
> @cvitolo <https://github.com/cvitolo> Thanks for your submission! Two
> things before I get reviewers:
>
>    - you do have a tests directory, but AFAICT they don't test the
>    functions in the package. Can you add at least a few tests that test the
>    functionality of the exported functions?
>    - You did check that you have a vignette, but the vignette is in the
>    parent directory, so not included with the package. Please move the
>    vignette dir into the package itself
>
