### 2016-08-10 #68 @cvitolo
@masalmon Thanks for spotting that. I have just removed all the old commented code and committed the changes.
