### 2016-08-10 #68 @masalmon
@cvitolo how long does it take for the vignette to build on your PC? It's taking ages on mine (but it seems to work, just slowly!).
