### 2016-08-10 #68 @masalmon
@cvitolo oh, maybe the examples are a bit too long for a vignette then? :smile: Maybe you could do the same for one of the countries only (e.g. Northern Ireland)?

Now I have the built vignette. :sparkles:
