### 2016-08-10 #68 @masalmon
# General
-	This a very cool package that allows to access data from a website without an API without having to parse the webpages. I’m very happy to see all the data it provides access to! Well done! :smile_cat:
-	I could build, test and check the package (on Windows). It behaves as expected.
-	Why is the package called rdefra and not ukair or rukair? Do you intend to add other DEFRA data sources? If not, ukair might be easier to understand but I know your package is on CRAN already so you might not want to change its name. ;-)
-	Devtools::spell_check finds two mistakes:
metres      stations.Rd:24
                   ourly       get1Hdata.Rd:18
-	 The repository is currently lacking a code of conduct (see https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#conduct ) which would help potential contributors see your repo is a friendly place. ;-)
-	In your DESCRIPTION file, you should use the Authors@R syntax https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#authorship (Authors@R is then parsed automatically when building the package)
-	The DESCRIPTION (and the vignette +  README) should include a link to UK air’s website.
-	Have you told people at UK air about your package? And can their website “resist” many people querying it (I have no experience about such issues)?

# Code
-	The function names use upper and lower case. It’d be easier to remember their names if they were all lower case. For ropenaq I was advised to prefixing every function with a package-specific string, which I really like now. For instance in your case the functions could be ukair_catalogue, ukair_eastingnorthing (or something easier such as ukair_impute_geo?), ukair_ site, ukair_data. The reviewer that had told me this for ropenaq had written that “This has the added benefit of playing very nicely with Rstudio, which will show all the package functions as completions when users type ukair_ and hit tab”.
-	Something that you could add are more thorough checks of the parameters supplied by the user. E.g. if the user gives “London” as a site ID in the get1Hdata, you could check “London” against the list of site IDs and then output an error message saying that “London” is not a siteID. It’ll be a bit of (boring) work for you but would make the package user-friendlier if the functions fail informatively. Currently if one asks for data about a site that doesn’t exist the URL doesn’t exist and you output a NULL object, maybe you could add a warning and describe the case in the function documentation?
-	When parsing data in get1Hdata, you could add a dependency to `lubridate` and output a datetime column instead of a Date and a time column (all measures are in the same timezone so you can use it)
-	Similarly in the stations data and in the catalogue you could output Start.Date and End.Date as a date.
-	Your function output very big data.frame. Why not output tbl_df/tibble instead so that they’re easier to print? For instance the end of get1Hdata would become `return(tibble::as_tibble(df))`. If you switch from read.csv to readr::read_csv (well that means another dependency) you’d get a tbl_df directly, I’m not sure what’s best.
-	For the measures, is it possible to get the corresponding units? It would be very useful information ;-)
-	The EastingNorthing function returns “A named vector containing Easting and Northing coordinates”. Why not a data.frame? It might be even easier if the function took the whole catalogue as an input and gave the “newly populated” catalogue as an output.
-	You write
```r
    message("Please insert a valid year (or sequence of years).")
    stop
```
You should write
```r
    stop("Please insert a valid year (or sequence of years).")
```
-	If you add dplyr as a dependency you could use standard evaluation, e.g.
```r
    df <- df[, -col2rm]
    df$site_id <- site_id
`` `
Would become
```r
df <- dplyr::select_( df, quote(-col2rm))
df <- dplyr::mutate_(df, lon = lazyeval::interp(~site))
```
Or possibly something easier. I’m not a standard evaluation expert, maybe someone has better arguments? Or is it even best practice for rOpenSci? There’s a vignette in dplyr about NSE https://cran.r-project.org/web/packages/dplyr/vignettes/nse.html and a chapter in Hadley’s advanced R book http://adv-r.had.co.nz/Computing-on-the-language.html
-	Does the cached version of stations include “guessed” coordinates?

# Dependencies
-	The packages you use in the vignette (raster, sp, leaflet) should be in Suggests: in DESCRIPTION.
-	Regarding XML parsing, rOpenSci guidelines mention “For parsing XML, if you only need to parse XML data, xml2 is a good option as it makes parsing easier. However, if you need to create XML (in addition to parsing it), use the XML package. The XML package is more low level than xml2, and allows finer manipulation, and may be favored by those more familiar with XML data.” I guess you can keep using XML?
-	The guidelines also mention “For http requests we strongly recommend using httr over RCurl.” Since you only use two RCurl functions, could you switch to httr? I think that `http_error` and `GET` are the functions you would use.
-	Regarding getURL, I don’t know if it behaves as GET but – as a nice ropenaq reviewer told me -- with httr::GET you can take a use  list of arguments to build the query. This can make constructing the call a bit cleaner, so that you don't have to paste0 them all together with &s. You can check the arguments near the top of the function, then construct a list of all arguments to pass to the queryargument of GET. For example:
```r
url <- "http://uk-air.defra.gov.uk/networks/find-sites?"
closed <- FALSE
pollutant <- 9999
arg_list <- list(closed = closed, pollutant = pollutant)
httr::GET (url, query = arg_list) # Any NULL elements of the list supplied to the query paramater are automatically dropped

```
-	I think you could replace plyr::rbind.fill by dplyr::rbind_all. I don’t know whether it’s better but dplyr is generally easier than plyr and I’ve read it’s faster.

# Readme and vignette
-	I know I have a big conflict of interest here but since you mention the openair package in the README, why not mention ropenaq as well? :-) OpenAQ has been providing access to the values of https://uk-air.defra.gov.uk/latest/currentlevels for a few months now and ropenaq allows to get that and data from other countries as well.
-	In the README/vignette, you could mention the temporal coverage of the data, and the parameters that are present in the data (PM2.5? O3?)
-	The vignette has no title and in the first line rdefra is written Rdefra.
-	If you want to keep the installation instructions in the vignette, then rather with eval = FALSE.
-	The vignette takes quite a while to build. Why not use smaller examples, e.g. only building the map for Northern Ireland? It might be useful to add  a chunk for not re-building the vignette on CRAN. I had to do this in the first chunk of ropenaq general vignette https://raw.githubusercontent.com/ropenscilabs/ropenaq/master/vignettes/Ropenaq-vignette.Rmd
-	The number of stations could change I guess, therefore you could replace the numbers in “There are 6563 stations with valid coordinates within the UK-AIR (Air Information Resource, blue circles) database, for 225 of them hourly data is available and their location is shown in the map below (red circle).” By `r nrow(IDstationHdata)`stations so that it changes if the number of stations changes.
-	I think you can filter the spatial data frame without over (and therefore with a dependency less? ) with “stationsSP[England, ]”.
-	At the end of your vignette you mention the README but users that installed the package from CRAN might not know where to find it. Why not had the code and explanations from the README in a eval=FALSE chunk?
-	In the vignette, you download data but you don’t show it. You could print the data.frame and maybe even add a plot?
After downloading the boundaries for the UK, maybe you should add file.remove(“GADM_2.8_GBR_adm1.rds”)?

# Documentation
-	You have generated the NAMESPACE by hand but you could actually rely on roxygen2 to do it for you. You can add #' @noRd to internal functions and #’ @export to exported functions. For dependencies you can add #’ @importFrom httr GET at the top of the file or write GET::httr each time you use  a function from another package (I always forget which way is the best one).
-	I *think* it’s best practice to also write the data documentation in a .R file that’ll be parsed by roxygen2.
-	The package lacks a top-level documentation for for ?package_name (see e.g. https://github.com/ropenscilabs/rnaturalearth/blob/master/R/rnaturalearth-package.r )
-	In the documentation of get1Hdata, you could explain how to input several years instead of one (the example is in the vignette but not in the function documentation).
-	The documentation of getSiteID should state where to find the IDs (from the catalogue I guess?).
-	The documentation of stations (and the vignette) could say how much one can expect the catalogue to change from one month to another for instance.
-	In the catalogue function documentation, could you explain how to find the country numbers / region_id/pollutants/etc? I see you had started doing this in details. You could even replace the numbers in input by something easier, e.g. the user would write pollutant=”o3” and you would translate it internally to the pollutant code.
-	In the catalogue function, you could describe more precisely what the three last arguments do. You could also skip them entirely if you think the non default mode doesn’t make much sense for users of your package.

-	Examples that shouldn’t be run should be packed in \dontrun{}, e.g.
#' @examples
#' \dontrun{
#'  get1Hdata("ABD", "2014")
#'}
#'
# Tests
-	Good job on now testing the package functions! Now, the tests are in 2 files. Best practice is to have one file for each function and to use context()at the top, e.g. in test-Catalogue.R the first line would be context(“catalogue function”).
-	Why not add a codecov.io badge to your repository? See https://github.com/jimhester/covr
-	Moreover regarding tests, you should add testthat::skip_on_cran before tests that might take a long time to run (the tests will still run on Travis and locally)

I don’t know how much time doing this took me, I had parallel tasks and I won’t count the vignette building, ahah. I’d say a few hours? I’m happy to help if you have any question!
