### 2016-08-10 #68 @sckott
thanks @masalmon !  can you give estimate of time it took to do review?
