### 2016-08-10 #68 @masalmon
@sckott not sure, see above, 2-3hours?
