### 2016-08-12 #68 @cvitolo
@masalmon Thanks a lot for your prompt and thorough review. I have already simplified the vignette. I will work on the other changes once I receive the second review. Many thanks again!!!
