### 2016-08-12 #68 @haozhu233
First, I really like the idea of this package as it provides R users an relative easy way to access the UK-AIR data, especially when the data source does not have a public API. Great job, @cvitolo! :)

Also, thank you, @masalmon, for providing such a thorough review for this package. You review covers many points I want to say. I'll try not to overlap our points too much. ;)

# Code Comments
- As @masalmon suggested, instead of using `getURL` and package `XML`, I would also recommend you to use `httr` and `xml2`.  `XML` is powerful but in this case, we can just take advantage of the convenience brought by `xml2`. Here is a short example of how to use that in the case of your `catalogue()` function.
```
library(httr)
library(xml2)

catalogue_fetch <- GET(
  url = "http://uk-air.defra.gov.uk",
  path = "networks/find-sites",
  query = list(
    site_name = "", pollutant = 9999,
    group_id = 9999, closed = "true",
    country_id = 9999, region_id = 9999,
    location_type = 9999, search = "Search+Network",
    view = "advanced", action = "results")
  )

catalogue_content <- content(catalogue_fetch)

# Please see the second comment
catalogue_csv_link <- xml_find_first(catalogue_content, "//*[contains(@class,'bCSV')]")
catalogue_csv_link <- xml_attr(catalogue_csv_link, "href")
```
- I noticed that in many places you used the location of a `<p>` tag to locate the value your are looking for. For example, again in your `catalogue()` function, you used `'//*[@id="center-2col"]/p[3]/a'` as the xpath. I would recommend you try not to using things like `p[3]` in this package as the contents of a webpage may change at any time. They may add a `<p>` tag if they change the descriptions a little bit. Instead, you can try to use a class as a locator to target your csv link as I showed in the example above.
- Some functions to fix those missing metadata are really slow. When I was going through the sample codes in README, it seemed to take forever for `EastingNorthing()` to accomplish its query. So I run a check and I got
```
system.time(EN <- EastingNorthing(stationList[1:10]))
#   user  system elapsed
#  0.170   0.032   8.174
```
In the sample code, `stationList` is a vector with a length of 2763 and that basically means it will take around 40 minutes to finish that line of code. My question is that when the information you are trying to fetch here is relative static and small(they are just the location data for different sites), do you mind to host a pre-cached version of the data somewhere (maybe github)? You can then write a function that fetch data from one link instead of asking each end user to send thousands of queries to the DEFRA server.
- You could consider to add a HTML status checking step in case your `GET` or `getURL` function returns something other than what you expected. If users of your package entered something that is not acceptable by the query and the server returns you something else other than a status 200, the function should stop and return an error.
- In `EastingNorthing()` and `getSiteID()`, instead of using `do.call`, `rbind` and `lapply`, you may consider to use `sapply`. You can try `t(sapply(stationList[1:10], EastingNorthing_internal))` to see the results.
- You may like to use `paste0()` to replace `paste(sep="")`.
- In `get1Hdata1`, `length(as.list(years)) == 0` might be too weak to validate the input.
- In `get1Hdata1`, you created a variable `id` to track if each attempt succeed. Maybe you want to return `id` as an attribute or a message when something failed?
- In `get1Hdata_internal`, instead of using `if (url.exists(myURL)){df <- read.csv(myURL, skip = 4)[-c(1),]...}`, you may want to use something like the following because sometimes even if a url exists, it doesn't mean you can download a csv from that url.
```
status <- try(read.csv(myURL, skip = 4)[-c(1),])
if(class(status) == "try-error")return(NULL)
```
- You may want to limit your code line width to the 80-character standard code width.
- I also recommend you to add `dplyr`, `tibble` and `readr` as your package dependency.
- I also recommend you to use roxygen to manage your package NAMESPACE. To export your function, you can add a `@export`. To import some functions from a package, you can add `@importFrom RCurl getURL url.exists`.
- You can put the `tests` folder under the root directory.
- When you write you test cases, it would be better if you can write your `test_that` descriptions in a statement format. You should also try to keep them as brief as possible

The review process took me around 3 hours
