### 2016-08-12 #68 @sckott
@haozhu233 thanks for your review!

@cvitolo both reviews are in, continue conversation here and let us know if you have any questions.
