### 2016-08-17 #68 @masalmon
You've done a very nice job improving the package, @cvitolo! And thanks for your detailed answers.

# General

* It's not best practice at all as far as I know, just a suggestion. Your DESCRIPTION is clear enough anyway.

* I now realize I've pasted the spelling mistakes without thinking too much about US vs. UK English, sorry.

* Great.

* Thanks for including us in the DESCRIPTION, but please not this is not a guideline, only the Authors@R formatting is. ;-)

* Cool!

# Code

* Nice!

* I have said something wrong, `dplyr::rbind_all` should be  `dplyr::bind_rows` because `rbind_all` is deprecated, I had to change it to have the tests pass.

* Great! I tested `ukair_get_hourly_data()`. The error message is very informative.
I think you could add a few more (or rely on the documentation you've improved!), because I for instance I still get

```r
> View(ukair_catalogue(pollutant = 66666666))
Error in View : No metadata available for the specified query
```

where if pollutant could be checked against the allowed values instead of "wasting" a query to the website.

* Cool to now have `datetime`. Currently you don't input the timezone so `lubridate` assumes it's UTC, but it should be BST (at least I assume the website gives BST?). You'll just have to add `, tz = "Europe/London"` in the lubridate calls.

* For some stations one gets with `ukair_catalogue` the start and end date are missing, therefore `lubridate` gives a warning. Maybe you could use `suppressWarnings` around the `lubridate` call to avoid that. Before doing that maybe check the dates for which there are warnings are indeed missing, because maybe they're given in another format than the other dates? If they are, you can use `parse_date_time` with the two possible formats as arguments, it would know how to choose.

* Regarding the units, I think you can use a `\deqn` command (or something else?) to have the greek mu in the documentation, although I've no experience with it. It's good to have the unit in the output if needed now! Your package seems to follow snake case in most cases, so maybe the argument should be `keep_units` instead of `keepUnits`? And, would it be hard for `unit.1` to be called `unit.Nitrogen.dioxyde` for instance when it refers to the `Nitrogen.dioxyde`?
Now that I think of it, maybe the unit information could be outside the output table, as an attributes, i.e.

```r
attr(output, "units") <- units
```

where units is a `tibble` with two variables, the name of the column in `output` and the corresponding unit. This would help having less columns in the `output` and still gives the `units` information.
By the way, what is the status information you remove?

* Reg. `ukair_get_coordinates`, what I had in mind was that the user only uses it for filling in missing longitude and latitude for a subset of the catalogue, so the user will have to translate northing/easting anyway, so could the input be the catalogue with missing and not missing longitude/latitude, and the function would fill in the blanks, so that the user wouldn't need to worry about how to convert Northing to a latitude?

* True, you might need to add `lazyeval` for using `mutate_` from dplyr.

# Dependencies

Ok, great!

# Readme and vignette

* Thank you, and no problem, `ropenaq` isn't very famous. :-)

* When you knit the README, doesn't it perform the installation instructions? Doesn't it annoy you?

* I get such errors/warnings when compiling the vignette

```r
Error in file(file, "rt") : no se puede abrir la conexión
Además: Warning messages:
1:  3070 failed to parse.
2:  5 failed to parse.
3: In file(file, "rt") :
  cannot open URL 'https://uk-air.defra.gov.uk/data_files/site_data/character(0)_1972.csv': HTTP status was '404 Not Found'
No data available for station character(0)
Error in file(file, "rt") : no se puede abrir la conexión
Además: Warning message:
In file(file, "rt") :
  cannot open URL 'https://uk-air.defra.gov.uk/data_files/site_data/character(0)_1973.csv': HTTP status was '404 Not Found'
No data available for station character(0)
Error in file(file, "rt") : no se puede abrir la conexión

```
The popup for stations on the map is "character(0)", because apparently `ukair_get_site_id(as.character(stationsRW$UK.AIR.ID))` doesn't return IDs for all stations? Does it make sense?

# Documentation

* Great that you added the details. You're truly increasing the accessibility of UK AIR data! You should just add a new line between information for pollutant and group_id for instance, and maybe even add lists with bullet points for each one to make it more readable?

# Tests

* Oh then I don't know what is best practice, maybe @sckott can help here?

# Misc

I still think you could add a nice plot to your README or vignette to make the package even more attractive. ;-)
