### 2016-08-17 #68 @sckott
Thanks for making those changes @cvitolo !

A few things:

* regarding tests: yes, `context` is used to group tests, so that's right. I think it's okay the way you have arranged the tests. I do always have a separate test file for each function in the package just to make things easier to find, but it's not a hard and fast rule.
* Depends vs. Imports: You don't need to use `Depends` in your DESCRIPTION file unless you plan to export functions from pkgs listed in `Depends` to users of `rdefra`. Right now you have  https://github.com/kehraProject/r_rdefra/blob/master/rdefra/DESCRIPTION#L15 - I'd move `lubridate, tibble, httr, xml2, dplyr` to `Imports` - You have correctly imported just the fxns  you need from those packages, so that's awesome
