### 2016-08-18 #68 @haozhu233
Hi, @cvitolo, you really have done an amazing job improving this package. It looks a lot better now!

Here are a few extra comments:
* Regarding your xPath question, I think you did a very good job in `ukair_get_coordinates_internal()`. In `ukair_get_site_id_internal()`, maybe you can try `page_tab <- xml2::xml_find_first(page_content, "//a[text()='Pre-Formatted Data Files']")`? I don't know for sure because they may also change the text.
* You don't have to use `@importFrom` in every roxygen documentation. Instead, you can just call all the functions you want to import in your `rdefra-package.R` file.
* Since you have already provided a cached version of stations data, maybe you want to edit your README a little bit so people can start with making some plots with your cached data? Or you can change the names of the variables in the README from `stations` and `stationsNew` to `stations_raw` and `stations`. So people won't need to make any changes to the plotting codes even if they want to use the cached data directly.
