### 2016-08-25 #68 @sckott
@haozhu233  - hey there, it's been 17 days, please get your review in by Aug 29, thanks :smiley_cat: (ropensci-bot)
