### 2016-08-26 #68 @cvitolo
Just fixed the figure in the README (although I cannot get rid of ).
