### 2016-08-26 #68 @jennybc
@cvitolo Not sure exactly what your `` problem is. But this rings a faint bell for me and might mean you need to update `rmarkdown` or `knitr`. If I'm off-base, just disregard.
