### 2016-08-28 #68 @cvitolo
Thanks for the suggestion @jennybc! I have updated `rmarkdown` and `knitr` but that did not fix the problem. As a temporary workaround I manually edited the README.md file and removed the string .

I also noticed that the leaflet map is not rendered and on its place is the line , not sure what's going on there.
