### 2016-08-29 #68 @masalmon
@cvitolo great work! A few points

* There's no unit test for `ukair_get_coordinates`  from what I see on codecov.io?

* In `ukair_get_hourly_data` you could output a warning if the side ID is not in the cached version of the catalogue. For instance imagine I enter a wrong ID, this is what I get:

```r
> ukair_get_hourly_data(site_id = "London", years = 1999)
Error in file(file, "rt") : cannot open the connection
In addition: Warning message:
In file(file, "rt") :
  cannot open URL 'https://uk-air.defra.gov.uk/data_files/site_data/London_1999.csv': HTTP status was '404 Not Found'
No data available for station London
# A tibble: 0 x 0
```
but it's a detail, the documentation is clear.

* I don't understand "Unfortunately the new attribute is not preserved when calling the function. "? Do you mean when calling another function such as `ukair_get_hourly_data`? Moreover ` attr(output, "units")` is still in the example in the function documentation.

* Regarding your README problem, try replacing the output part in the head of the Rmd by
```
output:
  md_document:
    variant: markdown_github
```
in any case, in the README you cannot render Leaflet parts, you can only show a screenshot, but you can have them in the vignette. If you want a nicer looking page than the README for promoting the package, you could still create a Github page for the package (maybe too much effort! [here](https://ramnathv.github.io/rCharts/) is an example of such a page)/add this on your personal website.

Your README is a bit long, so if you want a table of contents at the beginning, you can write

```
output:
  md_document:
    variant: markdown_github
    toc: true
```

* Nice to now see plots in the README. For having m^3 as an equation you could write
```r
plot(zoo(x = df$Ozone, order.by = as.POSIXct(df$datetime)),
     main = '', xlab = '', ylab = expression(paste('Ozone concentration (',
                                                   mu, 'g/',m^3, ')')))
```
and (really only a suggestion!) why not show the daily mean instead of all the points?

I agree with you that the package has greatly improved, nice job!
