### 2016-08-29 #68 @masalmon
I couldn't find a solution for `eqn` either.
