### 2016-09-05 #68 @masalmon
* It is weird that the test fails on Travis. You can still let the test and add `skip_on_travis()` before it. There's also a `skip_on_appveyor` function in `testthat`. By the way, do you skip the tests on CRAN with `skip_on_cran`? Since most of your tests probably take long, you could skip all of them on CRAN, see e.g. https://github.com/ropenscilabs/opencage/blob/master/tests/testthat.R By the way do you think the time now required to build the vignette is ok for CRAN?

* Ok I understand.

* Was adding the `attr(output, "units")` the last thing you did before returning the object? I have observed one looses the attribute with e.g. `mutate` calls if I remember correctly. You'll notice I comment a lot on that but `ropenaq` functions all output lists of data frames so... :smile_cat:

* Cool!

* I don't see the table of contents?

* Nice! By the way you write "Highest concentrations seem to happen in late spring and at the beginning of summer. In order to check whether this happens every year, we can download multiple years of data and then compare them" which is interesting but then you don't comment on the multiple years so it sounds a bit mysterious ;-) I have tried to re-do the second figure with

```r
library("rdefra")
library("lubridate")
library("ggplot2")
library("dplyr")
years <- 2013:2015
df <- ukair_get_hourly_data('MY1', years)
df <- mutate(df, year = as.factor(year (datetime)),
             year_day = yday(datetime))

df %>%
  group_by(year, year_day) %>%
  summarize(ozone = mean(Ozone)) %>%
  ggplot() +
  geom_line(aes(year_day, ozone)) +
 geom_smooth() +
  facet_grid(year ~.) +
  ylab(expression(paste("Ozone concentration (", mu, "g/",m^3,")")))
```
but it's not really prettier. :smile: I think it's quite hard to see seasonality. Maybe try it with another pollutant?

For your last question, I have no idea ("side effect is that the repository name will change and the link to the old releases will be lost")... @sckott ?
