### 2016-09-05 #68 @masalmon
I have tried downloading more years and I see the seasonality now:

```r
library("rdefra")
library("ggplot2")
years <- 2000:2015
df <- ukair_get_hourly_data('MY1', years)
df <- mutate(df, year = as.factor(year (datetime)),
             year_day = yday(datetime))

+
  ylab(expression(paste("Ozone concentration (", mu, "g/",m^3,")")))

```

![ozone](https://cloud.githubusercontent.com/assets/8360597/18241764/2c17211a-7353-11e6-8f8c-3c4f97726084.png)

Maybe you could comment on the later years which do not have the "nice" spike?
