### 2016-09-05 #68 @cvitolo
## rdefra - response to reviewers (step 4)

### @masalmon

* I added the tests and then used `skip_on_travis` before them. I'm not using appveyor at the moment (since I can only get it to work if the package is in the root directory. I'll use `skip_on_appveyor` function in the future,if needed. I use `skip_on_cran` only for the test that retrieve hourly data. The vignette seems to build relatively quickly now.

* Yes, I added the attr(output, "units") just before returning the object, but that does not preserve the newly added attribute.

* Ops, I forgot to knit the README.Rmd before committing. You should now see the toc.

* I have modified the second ozone plot based on your suggestions and added few comments.
