### 2016-09-06 #68 @masalmon
* It seems so unpractical that Appveyor can only make the build if the package is in the root directory, and I know you've tried `cd` commands in the configuration file... Maybe before doing the change you could try to ask a few people more if they know more about Appveyor and subdirectories? Unless you don't mind moving the package?

* I still don't see the toc ;-) Why not try

```r
output:
  md_document:
    variant: markdown_github
    toc: true
```
as header?

* A few comments about the README:

** Maybe add a link to the parallel package the first time you mention it and write `parallel` so that its name might get clearer?

** "the user access" -> "the users access"

** Nice new plot with the boxplots! It just lacks a x-axis label "Month of the year".

** Your meta should include "Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms."

* Regarding the unit tests, when I do R CHECK I get errors with the easting/northing functions but when I do `devtools::test()` I don't get errors... by the way I have just noticed `http_error` should be `httr::http_error()` since `httr` isn't loaded. The actual easting/nothing error seems to stem from a dependency issue `stop("package rgdal is required for spTransform methods")`. 1)This probably means rgdal should be in Imports in the DESCRIPTION (for Travis and users). 2) Somehow `rgdal` could be loaded when doing `devtools::test` but not R CHECK. I quickly tested my own suggestion and it did the trick for R CHECK, so if you add `rgdal` as a dependency you can skip the `skip_on_travis`.

* Regarding the `units`, I think it got lost because you added with `ukair_get_hourly_data_internal` and then made further operations on the output of `ukair_get_hourly_data_internal`. This made me wonder, if you ask data for several years, how does the units data frame look like? I tested

```r
output <- ukair_get_hourly_data("ABD", 2000:2014, keep_units = TRUE)
```
and then saw that the units data frame has 117 lines, some of them without unit.
1) Why not keep only unique lines with units?
2) do you think it's possible that in some cases O3 will measured with one unit for one year and then another unit for another year? If this can happen, shouldn't the units data frame have a year column?

* By the way getting units from the website to the user meant a lot of work for you and is clearly an advantage of using the package, so maybe you should mention units are returned in the README and vignette?
