### 2016-09-06 #68 @sckott
hi @cvitolo for  your question

> I was thinking of moving the package to the root directory (now it's in a subfolder) as I noticed that that is the easiest way to get appveyor to work. The side effect is that the repository name will change and the link to the old releases will be lost ...

You can do this. If you change the repository name, any attempts at going to the old URL will be redirected to the new URL, And I +1 this move of the pkg to the root directory 👏  just makes things easier.
