### 2016-09-07 #68 @cvitolo
> It seems so unpractical that Appveyor can only make the build if the package is in the root directory, and I know you've tried cd commands in the configuration file... Maybe before doing the change you could try to ask a few people more if they know more about Appveyor and subdirectories? Unless you don't mind moving the package?

I totally agree with you,  it is not practical! I am going to ask for help on stackoverflow or contact appveyor support.

> I still don't see the toc ;-) Why not try
> output: md_document: variant: markdown_github toc: true
> as header?

Sorry about this. I found out about that the output can be set to `github_document` and that removed the annoying html tags but forgot to indent the line`toc: true` and the toc disappeared. I have just commited the change.

> A few comments about the README:
> ** Maybe add a link to the parallel package the first time you mention it and write parallel so that its name might get clearer?

Ok, that's done

> ** "the user access" -> "the users access"

Done

> ** Nice new plot with the boxplots! It just lacks a x-axis label "Month of the year".

Done

> ** Your meta should include "Please note that this project is released with a Contributor Code of Conduct. By participating in this project you agree to abide by its terms."

Done

> Regarding the unit tests, when I do R CHECK I get errors with the easting/northing functions but when I do devtools::test() I don't get errors... by the way I have just noticed http_error should be httr::http_error() since httr isn't loaded. The actual easting/nothing error seems to stem from a dependency issue stop("package rgdal is required for spTransform methods"). 1)This probably means rgdal should be in Imports in the DESCRIPTION (for Travis and users). 2) Somehow rgdal could be loaded when doing devtools::test but not R CHECK. I quickly tested my own suggestion and it did the trick for R CHECK, so if you add rgdal as a dependency you can skip the skip_on_travis.

I did try that but did not work for me :(
If I add rgdal to imports I get the following error:
checking for gdal-config... no
no
configure: error: gdal-config not found or not executable.
ERROR: configuration failed for package ‘rgdal’
* removing ‘/home/travis/R/Library/rgdal’
Error: Command failed (1)
Execution halted

You can look at the full log [here](https://travis-ci.org/kehraProject/r_rdefra/builds/157904988).

> Regarding the units, I think it got lost because you added with ukair_get_hourly_data_internal and then made further operations on the output of ukair_get_hourly_data_internal.

You are right! setting the attribute in `ukair_get_hourly_data` rather than in `ukair_get_hourly_data_internal` fixed the problem. Thank you so much! Units are back in the attribute list :)

> This made me wonder, if you ask data for several years, how does the units data frame look like? I tested
> output <- ukair_get_hourly_data("ABD", 2000:2014, keep_units = TRUE)
> and then saw that the units data frame has 117 lines, some of them without unit.
> 1) Why not keep only unique lines with units?
> 2) do you think it's possible that in some cases O3 will measured with one unit for one year and then another unit for another year? If this can happen, shouldn't the units data frame have a year column?

The table has now a year column. It is possible that, over time, sensors will be substituted with new ones (with higher resolution, therefore different units?). My guess is that if this happens, the new sensor gets a new SiteID. I can check with DEFRA. In the meantime I think it is safer to keep the full table. Also, I wouldn't remove rows with empty units, otherwise users won't know if the units are missing or the function failed to retrieve all the metadata.

> By the way getting units from the website to the user meant a lot of work for you and is clearly an advantage of using the package, so maybe you should mention units are returned in the README and vignette?

Done.

Thanks for all your suggestions! Again, your help was precious!
