### 2016-09-07 #68 @masalmon
So Travis can't install gdal, is that the issue?
