### 2016-09-07 #68 @masalmon
@cvitolo I see that `geojsonio` depends on `rgdal` too, see https://github.com/ropensci/geojsonio/blob/master/.travis.yml maybe helpful? @sckott maintains the package so he should know more than I :smile_cat:

Then regarding `rgdal` in your NAMESPACE maybe only part of `rgdal` is needed (you imported all I think?) -- maybe have a look at what `spTransform` needs exactly?
