### 2016-09-07 #68 @masalmon
I also wonder whether you should make `rgdal` a suggestion given it's a huge dependency and is used in one function only... I'm not sure what's best practice.
