### 2016-09-07 #68 @masalmon
An example of package that has `rgdal` a Suggests only is `mregions`. See https://github.com/ropenscilabs/mregions/blob/fb95e15cd197ce39f08364ac02f21ec136387e50/R/region_shp.R -> maybe you could do something similar? The code does `check4pkg("rgdal")` and the doc says " if \code{TRUE}, you need the \code{rgdal} package installed.". You'd still have to get `rgdal` to work on Travis to be able to test your function but then not every user would need `rgdal` to work (which can be quite difficult I believe?).
