### 2016-09-07 #68 @mdsumner
Travis can install a binary R package like rgdal, but you need this in your
travis.yml even if it's a secondary dependency:

sudo: required
r_binary_packages:
  - rgdal

Working example is here:
https://github.com/SWotherspoon/SGAT/blob/master/.travis.yml

Travis can also build GDAL from source, see Edzer's .yml here:

https://github.com/edzer/gstat/blob/master/.travis.yml

So, I guess that's three ways to do it if you include the apt-get system
installs used by Scott's geojsonio. :)

spTransform() is defined in sp, but it's only useable if rgdal is also
available. It's kind of related to needing the projection metadata in sp -
so you can read and write data completely - even if you don't do any
transformations. You can use proj4 as an alternative to rgdal, but it still
needs PROJ.4 in the system, and it's a bit more restricted than rgdal with
spTransform.

Cheers, Mike.

HTH!
