### 2016-09-07 #68 @mdsumner
I'll add that I'm also happy to help get travis / rgdal working if you have
problems, whatever you decide to use.

Generally, rgdal is easily available on Windows and Mac, but it's a bit
trickier for some on Linux just because it can be done so many ways.  For
that reason using proj4 instead wouldn't help much, it's just a "lighter"
dependency overall.
