### 2016-09-07 #68 @sckott
thanks for the help @mdsumner - @cvitolo i'd take his advice over mine when it comes to spatial pkgs :)
