### 2016-09-07 #68 @cvitolo
@sckott Thanks, I have renamed the repo.

@mdsumner Many thanks for the advice and the offer to help! I have taken the 'build gdal from source' approach and now rgdal installs no problem.

I think this rgdal issue was the last thing to address.
