### 2016-09-07 #68 @masalmon
Yes I don't have other remarks (hopefully all potential users will be able to install gdal or ask for help). Nice job, and welcome back from Travis hell 😉
