### 2016-09-07 #68 @sckott
Great work!  A few items before approving:

* cran checks want you to use their "canonical form" of the url for a package, so e.g., change `http://cran.rstudio.com/web/packages/rdefra/index.html` to `https://cran.r-project.org/package=rdefra` throughout the package (roxygen docs, README, vignette)

* I noticed something that could be architected a bit better to fail better for users. `ukair_get_coordinates()` allows a few different types of inputs to its first parameter. Instead of allowing anything to be passed in, which e.g., can result in

```r
ukair_get_coordinates(5)
#> Error in xml2::xml_find_all(page_content, "//*[contains(@id,'tab_info')]")[[2]] :
#>   subscript out of bounds
```

use e.g.,

```r
foo <- function(ids) {
  UseMethod("foo")
}

foo.default <- function(ids) {
  stop("no 'foo' method for ", class(ids), call. = FALSE)
}

foo.character <- function(ids) {
  # do stuff
  ids
}

foo.data.frame <- function(ids) {
  # do stuff
  ids
}
```

which gives you nice failure behavior, for example:

```r
foo(5)
#> Error: no 'foo' method for numeric
```

You could alternatively, do the checking for correct class type internally without doing the above example, but the above does make it easy and clear to say what happens for each input class

I thin the other exported functions in the package are fine
