### 2016-09-11 #68 @cvitolo
@sckott Thanks for having another look at the package!

I have made the changes you suggested:
- use  canonical form for package url
- implement methods for ukair_get_coordinates, this is very neat, thanks for the suggestion!
