### 2016-09-12 #68 @sckott
Looks good to me. Approved!

Your vignettes still need a few pieces:

* You need `rmarkdown` and `knitr` in `Suggests`
* You need an entry in your `DESCRIPTION` file of: `VignetteBuilder: knitr`

To move your repo:

- Transfer to the repo to `ropenscilabs` github account
- Make sure to change all links to the new account, `ropenscilabs` instead of `kehraProject`

Please do follow these guidelines moving forward with your package:

* Do include more details in your `NEWS.md` file than are currently included. E.g. `Minor changes` isn't specific enough. And link to issues in your github repo like `(#43)`, which on Releases pages will hyperlink to the issue. See e.g., https://github.com/ropensci/taxize/blob/master/NEWS.md
* Make sure to git tag each version sent to CRAN with the matching version - __AND__ put NEWS items for that version in release notes, e.g. see e.g., https://github.com/ropensci/taxize/releases/tag/v0.7.9

Are you interested in doing a guest blog post on our site? http://ropensci.org/tech-notes/  If so, I'll get in touch with more details
