### 2016-09-13 #68 @cvitolo
Great @sckott! I am trying to transfer the repo to ropenscilabs but I get this message:

> You don’t have admin rights to ropenscilabs
