### 2016-09-13 #68 @sckott
@cvitolo You should get an invitation via email invited to a team on `ropenscilabs` account, accept that, and try again

> Also, should I change all the links in the badges as well (e.g. travis, codecov)?

yes, Once the repo is transferred, we can then turn on travis and codecov for your repo

> Happy to do a guest blog post, please send me more details.

I'll send an email to you
